requirements.txt
pyyaml
requests-oauthlib
python-dateutil
sphinx
